Tests: Determine all cruizlib modules to test for GUI imports
But exclude the workers api modules, since they import Conan packages, into the same process
and will infect all subsequent tests.
