+++ src/cruizlib/multiprocessingmessagequeuetype.py
@@ -3,9 +3,18 @@
 """Type annotation for multiprocssing Queues on Messages."""
 
 import multiprocessing
+import multiprocessing.queues
 
 from cruizlib.interop.message import Message
 
-# pylint: disable=unsubscriptable-object
-MultiProcessingMessageQueueType = multiprocessing.Queue[Message]
-MultiProcessingStringJoinableQueueType = multiprocessing.JoinableQueue[str]
+try:
+    # multiprocessing.Queue is a function rather than a class, so it does not satisfy
+    # the usual static type rules
+    # multiprocessing.queues is undocumented in more recent Pythons
+    # see https://github.com/python/cpython/issues/99509#issuecomment-1742069772
+    MultiProcessingMessageQueueType = multiprocessing.queues.Queue[Message]
+    MultiProcessingStringJoinableQueueType = multiprocessing.queues.JoinableQueue[str]
+except TypeError:
+    # pylint: disable=unsubscriptable-object
+    MultiProcessingMessageQueueType = multiprocessing.Queue[Message]
+    MultiProcessingStringJoinableQueueType = multiprocessing.JoinableQueue[str]
